requirements.txt
gspread
pyinstaller
Pillow